requirements_audit.txt
Flask>=3.0.0
Jinja2>=3.1.0
PyYAML>=6.0
